requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
html2markdown==0.1.7
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
readability-lxml==0.8.1
pydantic==2.4.2
aiohttp==3.8.6
markdown==3.5
lxml[html_clean]
# Alternative si l'option ci-dessus ne fonctionne pas
# lxml_html_clean 